Back-end/requirements.txt
flask==1.0.3
gunicorn==19.9.0
requests==2.22.0
jsonpickle==1.2
pyodbc==4.0.26
urllib3==1.25.3
urlparse3==1.1
faker==4.9.0
google-cloud-pubsub==2.1.0
pymongo==3.11.2
dnspython==2.0.0
Flask-Cors==3.0.9
